drop `default` channel concept
See discussion in https://github.com/indieweb/microsub/issues/16

Closes https://github.com/aaronpk/Aperture/issues/16

Creates an initial default channel when a new user is created, and allows that channel to be deleted. Now only the `notifications` channel can't be deleted.

The `notifications` channel cannot be moved or renamed, and is always first in the list.