add channel filters
closes #27

if any filters are set on the channel, checks the filters before adding to the channel. for both websub and adding existing entries to a new channel.

if any existing entries are added to a channel, or the first time a source is added to a channel, entries are marked as read.